first CRUD controller with MediatR

+++ MedHub/MedHub.Application/Handlers/DeleteAllergenCommandHandler.cs
@@ -1,13 +1,14 @@
 ﻿using MedHub.Application.Commands;
 using MedHub.Application.Mappers;
-using MedHub.Application.Response;
+using MedHub.Application.DTOs;
 using MedHub.Core.Entities;
 using MedHub.Core.Repositories.Base;
 using MediatR;
+using MedHub.Application.Helpers;
 
 namespace MedHub.Application.Handlers
 {
-    public class DeleteAllergenCommandHandler : IRequestHandler<DeleteAllergenCommand, AllergenResponse>
+    public class DeleteAllergenCommandHandler : IRequestHandler<DeleteAllergenCommand, Response<AllergenDto>>
     {
         private readonly IRepository<Allergen> repository;
 
@@ -15,11 +16,24 @@ namespace MedHub.Application.Handlers
         {
             this.repository = repository;
         }
-        public async Task<AllergenResponse> Handle(DeleteAllergenCommand request, CancellationToken cancellationToken)
+        public async Task<Response<AllergenDto>> Handle(DeleteAllergenCommand request, CancellationToken cancellationToken)
         {
-            var allergenEntity = await repository.GetByIdAsync(request.AllergenId);
-            await repository.Delete(allergenEntity);
-            return MedHubMapper.Mapper.Map<AllergenResponse>(allergenEntity);
+            var allergenEntity = await repository.FindFirst(allergen => allergen.Id == request.AllergenId);
+            if (allergenEntity == null)
+            {
+                return Response<AllergenDto>.Create(OperationState.NotFound);
+            }
+
+            repository.Delete(allergenEntity);
+            await repository.SaveChangesAsync();
+
+            var allergenDto = MedHubMapper.Mapper.Map<AllergenDto>(allergenEntity);
+            if (allergenDto == null)
+            {
+                return Response<AllergenDto>.Create(OperationState.MappingError, "An error occured while mapping object of type Allergen to AllergenDto.");
+            }
+
+            return Response<AllergenDto>.Create(OperationState.Done, allergenDto);
         }
     }
 }
